Move the `TaxRate` component to the `pages/settings/setup-tax-rate` directory.
