+++ js/src/pages/settings/setup-tax-rate/tax-rate.js
